+++ services/ske/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.11.0 (2024-03-28)
+
+- Waiters for async operation `StartCredentialsRotationWaitHandler` and `CompleteCredentialsRotationWaitHandler`
+- [Usage example](https://github.com/stackitcloud/stackit-sdk-go/tree/main/examples/ske)
+
 ## v0.10.1 (2024-02-28)
 
 - Update `core` to [`v0.10.0`](../../core/CHANGELOG.md#v0100-2024-02-27)